+++ packages/diblob-visualizer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @speajus/diblob-visualizer
 
+## 0.3.0
+
+### Minor Changes
+
+- b487df1: Added metadata ability
+
+### Patch Changes
+
+- d4b9fd7: Added embeded stuff help
+- Updated dependencies [d4b9fd7]
+- Updated dependencies [b487df1]
+  - @speajus/diblob@0.3.0
+
 ## 0.2.2
 
 ### Patch Changes